# voicescribe-api-gateway/openapi.yaml
openapi: 3.1.0
info:
  title: VoiceScribe API Gateway
  version: 0.1.0
  description: |
    Single public entry point for VoiceScribe AI.
    Authentication via JWT (Free Tier) or API Key (PRO/Enterprise).

servers:
  - url: http://localhost:8000
    description: Local development

paths:
  /v1/auth/login:
    post:
      summary: Login (Free Tier)
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many failed attempts
          headers:
            Retry-After:
              schema:
                type: integer

  /v1/auth/refresh:
    post:
      summary: Refresh access token
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Invalid or expired refresh token

  /v1/auth/logout:
    post:
      summary: Logout (invalidate refresh token)
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Logged out

  /v1/transcribe:
    post:
      summary: Submit file for transcription
      operationId: transcribe
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '202':
          description: Job queued
          headers:
            X-Request-Id:
              schema:
                type: string
                format: uuid
            X-RateLimit-Limit:
              schema:
                type: integer
            X-RateLimit-Remaining:
              schema:
                type: integer
            X-RateLimit-Reset:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscribeResponse'
        '401':
          description: Missing or invalid authentication
        '422':
          description: File too long for Free Tier (> 30 min)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Quota exceeded or rate limit
          headers:
            Retry-After:
              schema:
                type: integer
        '500':
          description: Internal server error

  /v1/jobs:
    get:
      summary: List jobs for tenant
      operationId: listJobs
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'
        '401':
          description: Unauthorized

  /v1/jobs/{job_id}:
    get:
      summary: Get job by ID
      operationId: svc01_getJob
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '401':
          description: Unauthorized
        '404':
          description: Job not found

  /v1/jobs/{job_id}/download/{fmt}:
    get:
      summary: Download job result
      operationId: downloadJob
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: fmt
          in: path
          required: true
          schema:
            type: string
            enum: [txt, srt, json, docx, vtt]
      responses:
        '200':
          description: File download
        '401':
          description: Unauthorized
        '403':
          description: Format not available for tier (e.g. DOCX for Free)
        '404':
          description: Job not found

  /health:
    get:
      summary: Health check
      operationId: svc01_health
      responses:
        '200':
          description: Healthy

  /metrics:
    get:
      summary: Prometheus metrics
      operationId: svc01_metrics
      responses:
        '200':
          description: Prometheus text format

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token (Free Tier)
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for PRO/Enterprise (format vs_live_...)

  schemas:
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        detail:
          type: string
          nullable: true

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer

    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer

    TranscribeResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          default: QUEUED
        message:
          type: string

    JobStatus:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
        status:
          type: string
        tier_at_creation:
          type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        error_message:
          type: string
          nullable: true

    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/JobStatus'
        total:
          type: integer